+++ sdk/resourcemanager/containerservice/armcontainerservice/agentpools_client.go
@@ -373,7 +373,6 @@ func (client *AgentPoolsClient) getUpgradeProfileHandleResponse(resp *http.Respo
 }
 
 // NewListPager - Gets a list of agent pools in the specified managed cluster.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-07-01
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // resourceName - The name of the managed cluster resource.