+++ sdk/resourcemanager/containerservice/armcontainerservice/models.go
@@ -183,27 +183,6 @@ type AzureKeyVaultKms struct {
 	KeyVaultResourceID *string `json:"keyVaultResourceId,omitempty"`
 }
 
-// CloudError - An error response from the Container service.
-type CloudError struct {
-	// Details about the error.
-	Error *CloudErrorBody `json:"error,omitempty"`
-}
-
-// CloudErrorBody - An error response from the Container service.
-type CloudErrorBody struct {
-	// An identifier for the error. Codes are invariant and are intended to be consumed programmatically.
-	Code *string `json:"code,omitempty"`
-
-	// A list of additional details about the error.
-	Details []*CloudErrorBody `json:"details,omitempty"`
-
-	// A message describing the error, intended to be suitable for display in a user interface.
-	Message *string `json:"message,omitempty"`
-
-	// The target of the particular error. For example, the name of the property in error.
-	Target *string `json:"target,omitempty"`
-}
-
 // CommandResultProperties - The results of a run command
 type CommandResultProperties struct {
 	// READ-ONLY; The exit code of the command